Add ServiceLocator
I chosen to call it a ServiceLocator, because most code in JTL-Shop is
not prepared for usage with a real dependency injection container.
Where possible, the ServiceLocator can be used like dependency injection
container.
Furthermore, I think most plugin developers will feel more comfortable
with the ServiceLocator-Pattern as this is simpler to understand and to
implement.